All Events are Pulled
the get_all_events() function will now loop through every event page list.
The server will only give you a list of events 15 entries at a time,
so you have to loop through "@odata.nextLink" until you get all entries.

The process of downloading files and writing them to disk was supposed to be async,
but that's not possible with requests and an incorrect attempt at that was added long with
this commit