Powder Springs added
This was an unpleasant scrape. The most difficult part is that there is
no consistent date data. All of the dates are stored in the file titles,
so they need to be heavily sanitized. This also means
there's a chance that if the file has no date, it can't be sorted.