Added NovusAgenda API (Kennesaw)
Kennesaw used to use NovusAgenda up until 2023, but now uses civic plus.
NovusAgenda has no API, and requires selenium. In order to make sure
this is reproducable I use the selenium docker image and use a remote
WebDriver, and all the user will need is docker (and selenium).

NovusAgenda is also very slow, so this whole process takes quite a
while. All of the minutes files links are pulled and then passed to
FileOps for downloading afterward so we can take advantage of threading.